requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
PyYAML==6.0.2
